Donator.Entity add field totalCompletedTransactions. Also add the fie…

+++ backend/src/main/java/donation/main/service/TransactionService.java
@@ -119,7 +119,7 @@ public class TransactionService {
                     .getTotalDonations()
                     .add(transaction.getContributionAmount());
 
-            Long count = transaction.getDonator().getTotalCompletedTransactions() + 1;
+            Integer count = transaction.getDonator().getTotalCompletedTransactions() + 1;
 
             transaction.getDonator().setTotalCompletedTransactions(count);
             transaction.getDonator().setTotalDonations(amount);
